refactor: submitted writing custom hook
